refactor: move subtract_p back to field12 base package
