+++ sxt/field12/operation/BUILD
@@ -11,9 +11,9 @@ sxt_cc_component(
     ],
     deps = [
         "//sxt/base/field:arithmetic_utility",
-        "//sxt/base/field:subtract_p",
         "//sxt/base/macro:cuda_callable",
         "//sxt/field12/base:constants",
+        "//sxt/field12/base:subtract_p",
         "//sxt/field12/type:element",
     ],
 )
